requirements_production.txt
# Emotion-Aware Voice Intelligence Platform
# Production Dependencies

# Core Framework
streamlit>=1.28.0

# Data Processing & Analysis
numpy>=1.24.0
pandas>=2.0.0

# Visualization
plotly>=5.15.0

# Audio Processing
soundfile>=0.12.0

# Async & HTTP
aiohttp>=3.8.0
asyncio-throttle>=1.0.0

# Environment & Configuration
python-dotenv>=1.0.0

# Utilities
pathlib2>=2.3.0
tempfile-wrapper>=1.0.0

# Development & Debugging (optional)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0